Fix crash in MapTreeView
If the tree view is only notified when objects will be deleted, it will still return those objects as children when the tree view updates itself, leading to stale entries in the tree view. Therefore, it has to be notified only after the objects have been deleted. Since we still need the parents to update the tree view, the objectsWereDeletedNotifier now takes an ObjectParentList instead of an ObjectList.,