Fix loading of texture rotation values on valve 220 maps.
The 'rotation' value in a 220 format brush is already incorporated into the texture vectors.
The old code was loading the texture vectors from the 220 format brush, and then rotating them again by the rotation value in the map file.

With this change, instead of rotating the texture vectors by the rotation value stored in the map, we load the texture vectors as-is, then copy the rotation value stored in the map into the BrushFaceAttribs as the inital value for the rotation.