Don't detach or delete nodes while merging them into the world.
When pasting nodes which were previously copied, the editor tries to merge the pasted
nodes into the world to avoid duplicate layers and a duplicate worldspawn entity. In this
process, some of the pasted nodes must be detached from their parents, and some must be
deleted because they will never be added to the world. Previously, this was done during
a visit to a pasted node. But since this will modify the structure being visited, it leads
to duplicate nodes in the result due to multiple visits to those nodes. Hence, nodes must
be detached and deleted after the iteration is done.