Disallow changing the selection in certain situations.
As an example, the uer cannot select brushes by clicking when the clip
tool is active, so it shouldn't be possible to deselect all brushes either.
While implementing this, I have also ensured that the event handlers in
MapFrame can also be called safely when the event source is not the menu.
This was necessary because I had added all menu key bindings to the
accelerator table of the map view to support one key bindings in
Ubuntu. Since there is no check whether an action can actually be
performed when the accelerator table is used, the event handlers
must perform the same checks as the UI update routine that enables
or disableds the menu items, too.