Add test case and fix bug.
The bug fix essentially consists of healing short edges after clipping
each plane into the brush while it is being built. Originally, this was
done after all planes had been clipped, but this was not enough as short
edges that are created during clipping may lead to bugs later on.