Reworking Brush::doCanMoveVertices (#1585)
Rework Brush::doCanMoveVertices so that it is more accurate in deciding which moves to reject and which to allow. Ensure that edge and face moves are only allowed if all edges / faces being moved still exist after the move was performed. Add appropriate test cases to cover the new behavior.