1404: Add simple AABB rebalancing based on node bounds volume.
When a node is out of balance (left height differs from right height by more than 1), then we remove a leaf from the higher subtree and insert it into the lower subtree. Thereby we select the leaf which would increase the bounds of the lower subtree the least among all leafs in the higher subtree.