2491: Don't crash on brushes with colinear face points. (#2522)
* 2491: Add failing test.

* 2491: Use smart pointers for tex coord system.

Since the BrushFace constructor may throw, we must make it exception safe. Using a smart pointer for the tex coord system is the easiest way to do this. While I was at it, I also changed the snapshots to use smart pointers.

* 2491: Use the same test for colinear face points in the parser as in the brush face itself.

The parser was using a different test than the BrushFace constructor to check the face points. In the brush attached to the issue, the check in the parser would pass, but the constructor check would fail, causing the issue.

* 2491: Ignore unused variable.