Added entity definition for testplayerstart (#2963)
testplayerstart is used for testing parts of your map quickly without having to noclip over to it every time you recompile. Instead of having to move your info_player_start around or copy it (which makes is a bit of a hassle because of entity edicts determining which one will be used), you can just create a testplayerstart somewhere and delete it when you're done. No muss, no fuss.