2816: Improve handling of missing textures and shaders (#2970)
* 2816: fall back to MATERIAL_NAME

If an ASE file does not contain a BITMAP directive, it falls back to the
MATERIAL_NAME directive to find the material path.

* 2816: package default texture with TB

The default texture __TB_empty.png is now packaged with TB itself
instead of individual game configurations (although it can be overridden
there). There is a new function loadDefaultTexture to load it from a
file system in ResourceUtils. The default texture is loaded automatically
when texture loading fails.