Fix multiple issues in CompilationRunToolTaskRunner (#2988)
There were multiple issues in the runner. First, it does not emit the
start() signal, and second, it immediately deletes the QProcess instance
when it finishes or when an error occurs. This commit changes it so that
the instance is only deleted when the runner itself gets deleted. Since
the runner itself is a QObject, it is not necessary to manually delete
the process, so the unique_ptr which managed the process is removed.

This commit also adds a test for the runner, but it's possible that this
test will not work on CI since it pops up the welcome window for some
reason.