3082: fix target/targetname editor immediately closing in some cases on Linux, when typing a letter to open the cell editor
Briefly, when typing the first letter of the text you want to enter to open the cell editor,
when setEditorData() runs, the letter has not been inserted into the QLineEdit yet.
Opening the completion popup and then typing the letter causes the editor to close, which is
issue #3082 and quite annoying. Only happens on Linux.

Fixes #3082