Make Brush::node accessor unavailable
In order to remove the Brush::m_node pointer, we need to get rid of all
code that depends on having this pointer. This commit makes the pointer
inaccessible to client code and updates the client code to not depend
on having this pointer available. Mostly, this requires using
BrushFaceHandle instead of BrushFace in client code, as the handle
contains a pointer to the brush node that contains a brush face.