Use BrushFace like a value type, part 1
In this commit, we make Brush take its brush faces by value and change
all places that create brush faces to create them on the stack.

Internally, Brush still stores the faces on the heap, and it returns
brush face pointers to its clients. This will be changed in the next
commit.