3229: hide unused surface and/or content flags in the flags editor UI (#3279)
* hide unused surface and/or content flags in the flags editor UI

Cf. issue #3229

Store a flag's bit value in the flag object. Then we can store the list of flags sparsely, omitting any unused flags.

The flagChanged signal now communicates both the flag's index in a list of flags as well as the flag's bit value, since some consumers care about the former and some care about the latter.

Unused flags are marked by the new "unused" property. If this property is specified then the flag is unused and no other properties need be specified. This new "unused" property is available in version 4 of the game config format. Version 3 is still supported.

In the editor UI, code that used to infer the flag's bit value from its index will now instead look at its stored bit value.

If there are no valid surface flags, don't show the surface value editor or the surface flags editor.

If there are no valid content flags, don't show the content flags editor.

* doc updates for config version 4

* docs updates for smart tags

* update game configs to version 4

If there are flags defined as "name":"unused", change them to "unused":true so they will be hidden from the flag editor UI.

Note that there may be a lot of other flags that are ACTUALLY unused but this change will not try to address that.

* more tweaks for smart tags docs

* enable -Wsign-conversion on gcc

* 3229: remove redundant ChangeBrushFaceAttributesRequest methods

* Revert "enable -Wsign-conversion on gcc"

This reverts commit f3f0c6d880b78c1071a469a6508bf4f8b55117e2.

* more docs fixes

* Add docs to FlagsEditor::flagsChanged

Co-authored-by: Eric Wasylishen <ewasylishen@gmail.com>