surfaceparm smart tag improvements
In version 4, accept a list for "pattern" value of a surfaceparm smart tag. The tag will match against any shader that has any of the listed params. A single value (as in version 3) is still accepted.

Using a set to store this list to make it easy to quickly find intersection with the set of surfaceparms.

Also support enabling "turn selection into" for surfaceparm smart tags. Factored some common code with the texture name smart tag matcher into a superclass.

Note that in TagMatcher.h the public-ness of some methods like enable() has been changed to match the access type of the TagMatcher methods they are overriding.