3288: Surfaceparm smart tag improvements (#3295)
* surfaceparm smart tag improvements

In version 4, accept a list for "pattern" value of a surfaceparm smart tag. The tag will match against any shader that has any of the listed params. A single value (as in version 3) is still accepted.

Using a set to store this list to make it easy to quickly find intersection with the set of surfaceparms.

Also support enabling "turn selection into" for surfaceparm smart tags. Factored some common code with the texture name smart tag matcher into a superclass.

Note that in TagMatcher.h the public-ness of some methods like enable() has been changed to match the access type of the TagMatcher methods they are overriding.

* more tag matching tests

Because of refactoring in the texture tag matcher, setting up a brush for those tests now requires setting textures on faces (not just creating a brush with a texture name).

Add test for texture pattern-based matching.

Add test for multiple surfaceparams matching.

Add test for enabling surfaceparam tag.

* preincrements to make Win build happy

* use kdl::vector_set instad of std::set, use std::begin over .begin()

* restore ability to tag match on names of invalid/missing textures

* allow matching against entire texture path

The current pattern matching had an undoc'd behavior where it would match the pattern only against the final slash-separated segment of a texture name.

However if the pattern itself contains a slash, we should match it against the entire texture name.

Co-authored-by: Eric Wasylishen <ewasylishen@gmail.com>