2983: Use result type for BrushFace::create
This modification is a blueprint for future commits where result types
are introduced. We introduce the type and handle it at all of the call
sites by throwing an exception, and we mark the throw with a comment.
Then we transform each of these call sites to use result types or handle
the error until no more such comments remain.