3410: Make texture collections immutable
When TextureCollection becomes a value type, we must take extra care to
ensure that texture pointers doin't start dangling because the
referenced texture is moved to a different memory location. This is
mostly ensured by std::vector's move semantics, however, there is the
case of reallocation when textures are added to an existing texture
collection. This commit avoids that case by making it impossible to
modify a texture collection's texture vector after it was created.