3421: generate more precise diffs in EntityAttributeModel
the previous code only handled a few cases and it would give up and reset the model
if rows were both added and removed. This change fixes the bug where a
selected cell (e.g. "target") would get deselected if you clicked on a
different entity with a different classname that still had a "target" cell.