3523: Add facilities to visit nodes with lambdas
This commit adds two new member functions to Node which accept a lambda
as a visitor. We decided to omit all the existing functionality to
control iteration and recursion; instead this will be implemented in the
lambdas and at the call sites. This allows us to remove the facilities
to cancel a visitor or to stop recursion.

Two new functions are now called `acceptLambda`, these will be renamed
to just `accept` when all existing visitors are replaced by lambdas and
the existing facilities are removed.