Fix compiler launch scripts
When using ccache under macOS or Linux, we generate scripts that launch
the system compilers. This is due to a limitiation in Xcode's build system
that prevents it from using ccache, but we apply this to all generators on
macOS and Linux anyway.

The scripts contain a syntax error that prevents them from working on Linux.