3619: Rename AttributeDefinition files
In this step, we only rename the files to keep the following commit, in
which we rename the associated classes as well, a bit cleaner.

3619: Rename AttributeDefinition files

In this step, we only rename the files to keep the following commit, in
which we rename the associated classes as well, a bit cleaner.