3043: Inline the gtest compatibility macros (#3643)
This commit removes all GTest compatibility macros that were introduced when we switched to the Catch testing framework. This consists of mostly mechanical changes, but in some instances, we replaced the existing assertions with more advanced Catch features such as the `GENERATE` macro. We also refactored test helper functions, for example changing the parameter order to take the expectations as the last argument.