558: Add linking facilities to GroupNode
We introduce the notion of linked groups here and make sure that a group
is always linkable to itself upon creation or cloning. Cloning a group
node creates a new link set, so the clone cannot be linked to the
original.

To store the link set shared data, we introduce the SharedData struct
which is managed via a shared pointer.