3656: Refactor map parser / reader classes to avoid using ModelFactory
MapParser had some unused support for auto detecting map formats. Even
though we want to have this feature in the future, I have removed it
here to simplify this refactoring. This means that the map format is
passed to the parser during its creation.

To support on the fly conversion of parsed faces, I have opted to pass
two map formats to the MapParser constructor: one for the format of the
source material (the string to parse) and one for the format of the
faces to create. IMO this makes this distinction a lot clearer.

Furthermore I have removed all use of ModelFactory from these classes.