Remove kdl::result factory functions
We had introduced factory functions for kdl::result, but using these
requires repeating a lot of typing information that is not strictly
necessary. This commit replaces the factory functions with constructors
to construct result types from a value, which is either the success
value or an error. These constructors are not explicit on purpose so
that we can return a value from a function that returns a result type.

Furthermore, I'm adding a converting constructor that can construct a
result from a compatible result. A result type B is compatible to a
result type A if B has the same value type and a subset of the error
types of A.