Replace loops with nested result handling with for_each_result
This works really great in some places, e.g. BrushBuilder.cpp, and ok
in other places. There is still some a problem with juggling too many
types in some algorithms, and it might result in code that is hard to
read for someone who isn't used to this style.

On the other hand, the algorithms where I replaced the manual loops get
a really nice pipelining structure because we can now just use and_then
to chain all stages of a transformation together.