3561: Take mouse position in double precision from Qt events
This doesn't actually fix the issue yet, because for some reason
QMouseEvent::localPos only returns integer values. We'll get to that in
the following commit.

3561: Take mouse position in double precision from Qt events

This doesn't actually fix the issue yet, because for some reason
QMouseEvent::localPos only returns integer values. We'll get to that in
the following commit.