3545: Remove DocumentCommand
Currently the decision of whether or not a command can update the
document's modification count is made statically depending on whether
or not the command extends DocumentCommand. We want to allow
SetLockStateCommand to make this decision at runtime, so we merge
DocumentCommand's functionality into UndoableCommand and make that take
a new boolean parameter that controls the new functionality.