document map bounds
Closes issue #3718

One thing to note is that the existing "main editing window" image MainWindow.png is old enough that it does NOT show the Map Properties widget. I didn't try to re-create this image because I don't know if there are particular concepts that you wanted to make sure were displayed in this image.