3743: Don't start an event loop during tests
CompilationRunnerTest starts an event loop in order to be able to use
QTimer to wait for a timeout. Starting an event loop when there is an
instance of QApplication leads to the TB welcome window popping up
during the test run, which is annoying (and slow).

This commit replaces QTimer with a mutex and a condition variable so
that we can use std thread primitives to wait for the runner to exit.