3725: Use vertex position matching to set face attributes after CSG merge
This change introduces a new method of setting face attributes after performing
a CSG convex merge operation. For each newly created face, we find the best
matching face of the source brushes. A face of a source brush receives a match
score based on the number of vertices it shares with the newly created face.
Ties are broken using the face normals.