2429: Cleanup: Get rid of linker warnings when game config has obj files
Since we have to add game files as source files on macOS in order to set
their package locations in the app bundle, the linker will try to link
any .obj model files in game configs. Since these files cannot be
linked, we get a lot of warnings when building on macOS. This change
tells cmake to treat game config files like header files, so it doesn't
try to link them.