3757: Update linked group nodes after setting properties to unprotected
After setting some properties to unprotected, we need to update linked
nodes because in some situations, setting a property to unprotected can
lead to it being replicated into the corresponding entities in linked
groups. Suppose we have two linked groups, each containing an entity.
One of the entities has some protected property that the other doesn't.
In this case, when the protected property is set to unprotected, it must
be replicated into the corresponding entity.