2429: Refactor MapReader
In this refactoring, we pushed more implementation details into the cpp
file. Other than that, the most important change is that we store all
infos collected during parsing in one vector that we then process in
parallel, creating all nodes in parallel as well. We then resolve any
implicit parent / child relationships and finally call the callbacks.