2429: Refactor map reader (#3772)
* 2429: Remove support for parsing extra attributes

In TB1, we used extra attributes to store information such as issue
filters and node visibility, but this has never been ported to TB2 and
there is no plan for supporting these features.

* 2429: Refactor MapReader

In this commit, we refactor MapReader with the goal of moving a lot of
implementation details into the cpp file as free functions. We hope that
this will make it a little bit easier to understand the code and to
extend it.

* 2429: Remove default layer properties from worldspawn after use

These properties should not remain in worldspawn after they have been
examined.

* 2429: Move entity properties after parsing to avoid unnecessary copies

* 2429: Refactor MapReader

In this refactoring, we pushed more implementation details into the cpp
file. Other than that, the most important change is that we store all
infos collected during parsing in one vector that we then process in
parallel, creating all nodes in parallel as well. We then resolve any
implicit parent / child relationships and finally call the callbacks.