3774: Close compile dialog when creating or opening a document
The compile dialog works on the compilation configuration for the
current game and this doesn't update when the game changes. When a new
document for a different game is created or loaded, then the compilation
profiles from the previous configuration will overwrite the compilation
profiles for the game of the new document. To prevent this, we close the
compile dialog when a document is created or loaded.