3688: Ignore unselected nodes when updating vertex handle manager
Since it is possible to modify unselected nodes in some cases, it can happen
that nodesWillChange / nodesDidChange notifiers get called with unselected
nodes. In the vertex tool, this will trigger the addition / addition of vertex
handles of unselected brushes. Adding handles for unselected brushes is
unexpected, but not fatal. On the other hand, removing handles for unselected
crashes in debug builds, where we assert that the handles to be removed were
previously added to the handle manager.