3781: Add first() and all() members to PickResult
These functions will replace the query() member function. To make the
transition easier, we extract helper functions firstHit and allHits
which are then used from HitQuery and PickResult in the corresponding
member functions. In the next commit, we stop using HitQuery altogether
in favor of using the new members of PickResult.

firstHit no longer provides a means to set an occluder hit type. It
turns out that every use of this function would always follow the same
logic: Every hit type is used as an occluder. Therefore, we can remove
this parameter and always assume this inside of the function.