3781: Refactor Notifier
This refactoring makes Notifier a little simpler, incl. using the
corresponding classes NotifyBefore, NotifyAfter, and such. It also fixes
some issues related to forwarding the notification arguments correctly
to notifiers.

We also change how observers are connected and disconnected. Previously,
to disconnect an observer, we had to pass the same observer to the
removeObserver function. This relied on comparing function pointers.
We now avoid this by returning a new object that represents one or
multiple connections from observers to notifiers. This object will
disconnect its connections when it falls out of scope, which allows us
to remove a lot of code that did this manually.

Since the return value of connectObserver is marked as nodiscard, the
compiler will warn us when we drop a connection immediately.

Unfortunately, it was not possible to split this refactoring into
multiple smaller steps in a natural way, so we did it in one big commit.

The need for this refactoring came up when working on #3781, where we
ran into errors due to the faulty forwarding of notification parameters.