3781: Make Notifier move-only
Notifier must be a move-only-type because otherwise not all observers
get disconnected properly if a notifier is copied by accident. This
reveals a bug where connecting one notifier to another copies the
observing notifier. To prevent this, we provide a separate overload to
Notifier::connect that takes the observing notifier by reference,
avoiding the copy.