2102: Refactor EL
This refactoring contains a number of cleanups and simplifications. The
main takeaway is that we now store the expression that generated a value
as a member of that value. To that end, we make both Value and
Expression trivial to copy by storing their actual data using
shared_ptr. To achieve this, we must make values and expressions
immutable.

We also extract all of the concrete expression types into a separate
compilation unit to improve source code organization.