3816: Allow non default constructible in vec_parallel_transform
vec_parallel_transform can only be used when the result of the transform
function is default constructible. This change wraps the result in an
optional and automatically unwraps it when returning the results. Since
optional is default constructible, this makes vec_parallel_transform
usable with non default constructible result types.