3848: Ensure that deleted notifiers are ignored when disconnecting
When NotifierConnection is destroyed, it attempts to disconnect all
stored connections. To ensure that it skips this for notifiers that have
already been destroyed, we introduce a new class NotifierState which is
managed via a shared pointer. NotifierState then only holds weak
pointers to the notifiers it needs to disconnect later, and if such a
weak pointer has already expired, the connection is no longer attempted
to be disconnected.