3909: Always update tool drag state in input state
This is necessary to correctly reflect that any tool is dragging in all
views, not just the one in which the drag is taking place. Otherwise, if
a drag is started in a 3D view, a 2D view would still have this set to
false, which could lead to errors.