Set fetch depth to unlimited to avoid vcpkg error on CI
The reason for this change is that vcpkg needs a full clone with all
history in order to use versioning for dependencies. There is no way to
use different fetch depths for git submodules and the host repository
in github's clone action, so we change to doing full clones for now

todo! check if we can avoid a full clone of TB's history