Check for property definitions when determining the rotation info
The rotation of an entity should be determined by its entity properties,
but in absence of such properties, we must inspect the property
definitions instead. The effect is that users can apply rotations to
entities that don't have corresponding entity properties such as
"angles" can still be rotated without adding such a property if the
entity has a property definition for an "angles" property.