Make edge healing more robust
The previous algorithm tried to be smart about removing invalid (short)
edges by not restarting the search when a short edge was removed. We
change it so that it starts over, which is less prone to errors in cases
where more than the short edge itself is removed. This avoids an
assertion at the cost of more work when edges are removed.