Correctly invert the dot product
To snap extruded faces to the grid, we use the dot product to convert a
distance along some direction into a distance along some other direction
(namely, an edge). Then we snap that direction and convert the snapped
distance back to a distance along the original direction - and here we
cannot just use the dot product, but have to properly invert the dot
product.