Prefer entity properties over definitions when determining rotation type
When determining the rotation type, we look at both the actual entity
properties and the entity property definitions from the FGD file with
equal priorities. This commit changes this behavior to prefer the actual
properties over the property definitions so that entities with more than
one defined angle-related properties (e.g. misc_model for Quake 3, which
has both a defined angle and a defined angles property) are still
displayed correctly. Before this change, the definition for the angles
property would override an actual angle property and so the entity
rotation would be computed wrongly.