Remove model format whitelist
With this commit, we make the model format whitelist optional. If still
present in a config file, it is ignored. We increase the game config
format version to 5, which makes means that older versions of TB cannot
load a migrated game config file, but newer versions can still load
versions 4 and 3.