Handle blank scale value
Instead of returning Undefined for missing entity properties, we revert
the changes introduced in 4981f19f and handle a blank scale value when
evaluating model expressions. This fixes a bug where accessing non
existing spawnflags in a model expression would not yield the expected
results since the spawnflags would no longer default to 0.