Specify transaction scopes
We need this information later to determine at which point to update
linked groups: If a one shot transaction is committed, we want to update
linked groups unless an enclosing transaction is also a one shot
transaction. If a long running transaction is committed, we don't need
to update linked groups because the nested one shot transactions will
already have updated them.