Extract UpdateLinkedGroupsCommandBase
We want to issue linked group updates as commands that can be collated
into their predecessors immediately. This way, we can block such an
update from happening depending on the state of the command processor.
These commands must be collated into their predecessor so that they
don't block further collation of succeeding commands.

We first extract the linked group update facilities into a separate
class that we can then reuse to implement an UpdateLinkedGroupsCommand.