allow use of MAP_DIR_PATH in tools
For issue #4010. This makes the code align with the documented behavior.

All this change does is bump MAP_DIR_PATH up in the "hierarchy" of variables,
taking it out of CompilationWorkDirVariables and into its parent class
CommonCompilationVariables which is used both for workdir and tool
interpolations.

This does mean that CompilationWorkDirVariables now is functionally identical
to CommonCompilationVariables, but it seems fine to leave them as two distinct
classes to make it easy to handle any future changes where they diverge again.