Enable short circuit evaluation of logical and case operators
Logical and and logical or and the case operator should be evaluated
with short circuiting. In particular, this means that lhs && rhs should
not evaluate the right hand side expression at all if lhs evaluates to
false, and likewise for lhs || rhs if lhs evaluates to true. For the
case operator, the right hand side after the -> should not be evaluated
if the left hand side evaluates to false.