Explicitly commit transactions
The Transaction helper class commits transactions automatically in its
destructor. This is convenient, but we want to update linked groups
as part of commiting transactions, and updatling linked groups can fail.
So we make committing an explicit function call, and in a later commit,
Transaction::commit will return a bool indicating whether updating
linked groups was successful.