Fix for parsing unrecognised FGD class properties (#4084)
Allow parsing of unknown FGD properties without parentheses.

@PointClass unknownprop base(Targetname) = my_entity : 

In the above example, unknownprop is not recognised by TrenchBroom, so FgdParser::skipClassProperty() is called to skip any parentheses. However, since the unknownprop token has already been consumed and no parentheses are present, the token "base" is encountered and skipped, and then the function exits since the parentheses depth is 0. Parsing subsequently fails since the parenthesis that is read next is unexpected.