4010: make MAP_DIR_PATH interpolation available for tools (#4098)
* allow use of MAP_DIR_PATH in tools

For issue #4010. This makes the code align with the documented behavior.

All this change does is bump MAP_DIR_PATH up in the "hierarchy" of variables,
taking it out of CompilationWorkDirVariables and into its parent class
CommonCompilationVariables which is used both for workdir and tool
interpolations.

This does mean that CompilationWorkDirVariables now is functionally identical
to CommonCompilationVariables, but it seems fine to leave them as two distinct
classes to make it easy to handle any future changes where they diverge again.

* add test for interpolation in compilation context

This test just checks that MAP_DIR_PATH and WORK_DIR_PATH can be interpolated
in context of compilation tools. Added because the ability to do so with
MAP_DIR_PATH is new.

BTW the change in header file arrangement comes from running clang-format.