Provide our own version of poly2tri
This library produces a bunch of linker warnings on macOS because it is
configured without symbol visibility settings on clang. TrenchBroom
itself and all other libraries are compiled with hidden symbol
visibility, so during linking, the conflicting settings lead to
annoying warnings.

This commit introduces our own version of the library. The only
difference is that we set CMake's policy 0063 to default to YES, which
instructs CMake to apply visibility settings to all types of libraries.