new "single target" checkbox for the Copy Files task
As discussed in issue #4047, having this option would address the limitations of the current "copy to directory" behavior in a way that tackles a couple of the affected usecases.

This would also address the documentation part of issue #4019 (or at least make it moot). As a way to provide a platform-independent tool for this task it is also work toward issue #1924 and issue #2033.

Note that this does NOT address config file versioning in any way (cf. the comments for issue #3887). If a config that includes a copy-files task with this option is somehow used with an older version of TB, parsing that config will fail with an "Unknown compilation task type" exception. However, existing config files that include a "Copy Files" task will still work with this code, since this new flag is optional (defaults to false).