Only cancel mouse drags if window deactivates
Some windows managers deactivate views when the mouse cursor leaves them
regardless of whether a drag is in progress. This would lead to mouse
drags being cancelled inadvertedly because we cancel mouse drags when
a map view or the UV view lose focus.

In this commit, we change it so that we cancel mouse drags when the
window is deactivated.