Implement animations for Assimp models, if present
This change detects if an Assimp model has skeletal animation
and creates a "frame" for each animation in the model. The user
can then map their framekey to a property (such as "sequence")
to preview the first frame of each sequence in the editor.

I addded this specifically for HL1 animation support, but this
uses Assimp's generic animation mechanisms, mostly by looking at
the assimp viewer code. I tested this on a few other static models
and they worked fine, however it is quite difficult to find
other animated models that Assimp would load, so it's not tested
for any other animated model formats.

When animations are not present in the Assimp scene, it will default
to using the same behaviour as it did before.

Contains a fix to detect when Assimp has loaded HL1 studio models
and handle its issue of rendering all submodels for each body
part, instead of just one submodel per body part. Supporting
selectable body parts is quite possible, but will need changes to
other parts of the application so extra keys can be passed through
to the asset loader.

Related to #1140