Refactor loading so that we load one single-textured surface per mesh, instead of one multi-textured surface
Assimp meshes always have one texture, so this is a valid change to make. Assimp supports skins by adding "alternative textures" to each mesh, which is why this is required.

We inline the logic from `computeMeshFaces` as we no longer need the vertex index offset since each mesh starts at index 0 in its respective surface.

We put the data into the `meshData` list because we need to compute all the vertex positions prior to `loadFrame` on the model, as we need the bounding box. So we pre-compute all the data, get the bounds, and then bind all the mesh data onto each surface after loading the frame.