Retry Cpack on macOS CI builds 3 times
For a while now, CPack has been failing occasionally on macOS CI builds.
Specifically, hdiutil fails to create a DMG image with the error message

  hdiutil: create failed - Resource busy

There are reports about such failures such as this github issue:

  https://github.com/actions/runner-images/issues/7522

However, this issue is about the macos-13 github runner, and we are
using the macos-12 runner. In our case, retrying the failing build
usually fixes it, so we now attempt to run CPack 3 times before giving
up.