Add a failing test to demonstrate an issue with proptected properties
The test shows that protected properties do not work when the structure
of the changed group changes. This is obvious when looking at the code,
which relies on the a parallel traversal of the cloned subtree and the
corresponding subtree to find the entity nodes that carry the original
values of any preserved properties.

The following commit will fix this by replacing the parallel traversal
by a simple hashmap where we can look up the corresponding nodes by
their link IDs.