Don't apply changes to faces linked to the source's containing group
Let F be a face in a brush B that is in a linked group G. Let X be
a face in a brush Y that is in a linked group G' such that G and G'
have the same link ID.

Then applying a change to face X will change G', and that change will
be replicated into group G. Since all nodes in G are replaced, node
pointers and brush face references are no longer valid. This leads to
inconsistencies and bugs.

We can avoid this problem by applying the change not to face X, but to
the face in G that X is implicitly linked with. That way, group G gets
replicated and any pointers or face references to objects in G are still
valid afterwards.