Mount wad files directly under textures/
Wad files used to be mounted under textures/<wadname>. This was useful
because it allows to filter the wad files in the texture browser
settings, and it also allowed us to show overridden textures in the
browser. At the same time, it complicates handling of texture names
because the texture names cannot be extracted from the file paths in a
uniform matter across different games.

We change this to simplify texture loading, in preparation for the new
material system.