Store user config files in a suitable location
The user config files were stored in a path derived from the game's
display name as configured in the game config. This name could contain
invalid characters which would crash TB when saving user config files.

This commit changes this behavior to store these files in a directory
named after the directory that contains the game config - which must be
a valid name. Existing config files are migrated automatically.

As a consequence, users switching between this version of TB and older
versions might have to manually undo the migration to keep their user
ocnfigurations intact.