Disable ccache on CI
Caching ccache in the action cache thrashes vcpkg's cache, and that defeats the
purpose because the builds actually slow down. We leave the changes in this PR
because using ninja on Windows has other benefits, too. If we get better control
over github's action cache, we can re-enable this feature by reverting this
commit.