Cleanup: Rename TargetDestination to LinkSource and TargetSource to LinkTarget
In FGD files, target_destination marks the property that holds the name of the
link target, while target_source marks the property that holds the name to
target. In TB, we call these properties the link source and link target,
respectively, so we rename the property value types to reflect TB's internal
semantics.