Add new color types and functions
This commit replaces the existing Color type with a new type that internally
stores colors as a variant of one of four types. We allow colors with RGB or
RGBA components, and float or byte as the component types. Unfortunately we have
to adapt all clients of the new type in one big commit.