Run windeployqt for dump-shortcuts
This solves an issue when compiling and Qt is not on the path (which
isn't required for the main app anymore). dump-shortcuts attempts to
run with every build and fails due to the missing binaries, causing
minor annoyances due to the app compiling successfully but the overall
build failing.

The only change to behaviour from the main app's windeployqt is that
translations are never copied, as they are not needed to dump shortcuts.